Photo reel
Not done but started, needs work, need to use something htats not Javascript for this by the looks of it, as java script can read a file fine but when it comes to reading entire directories it suffers. Quick look online shows PHP and Ajax are viable for the task I need.